requirements.txt
Flask==3.0.2
Flask-Cors==6.0.1
requests==2.31.0
boto3==1.38.36
PyJWT==2.10.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.1
Werkzeug==3.0.3
